Clean up some more code generation
